Add current line highlight
