+++ AuroraEditor/Base/Documents/UI/WorkspaceCodeFileView.swift
@@ -240,25 +240,8 @@ public struct AECodeView: View {
             }
 
             // Minimap
-            ZStack(alignment: .topLeading) {
-                Color.black.opacity(0.2)
-//                Color.gray.opacity(0.5)
-//                    .frame(height: minimapLineHeight)
-//                    .offset(y: CGFloat(sharedObjects.caretPos.line) * minimapLineHeight)
-                ForEach(attributedTextItems) { textItem in
-                    if let color = textItem.attributes[.foregroundColor] as? NSColor {
-                        Color(nsColor: color)
-                            .frame(width: CGFloat(textItem.text.count) * minimapMultiplier,
-                                   height: minimapLineHeight)
-                            .offset(x: CGFloat(textItem.charactersFromStart) * minimapMultiplier,
-                                    y: CGFloat(textItem.lineNumber) * (minimapLineHeight * 1.5))
-                    }
-                }
-            }
+            MinimapView(attributedTextItems: $attributedTextItems)
             .frame(width: 150)
         }
     }
-
-    let minimapMultiplier: CGFloat = 0.8
-    let minimapLineHeight: CGFloat = 2
 }
